edit 6-size.c

+++ 0x00-hello_world/6-size.c
@@ -16,11 +16,11 @@ int main(void)
 	char c;
 	float f;
 
-	printf("Size of an int: %ld byte(s)\n", (long int) sizeof(i));
-	printf("Size of a char: %ld byte(s)\n", (long int) sizeof(c));
-	printf("Size of a long int: %ld byte(s)\n", (long int) sizeof(li));
-	printf("Size of a long long int: %ld byte(s)\n", (long int) sizeof(lli));
-	printf("Size of a float: %ld byte(s)\n", (long int) sizeof(f));
+	printf("Size of an int: %ld byte(s)\n", sizeof(i));
+	printf("Size of a char: %ld byte(s)\n", sizeof(c));
+	printf("Size of a long int: %ld byte(s)\n", sizeof(li));
+	printf("Size of a long long int: %ld byte(s)\n", sizeof(lli));
+	printf("Size of a float: %ld byte(s)\n", sizeof(f));
 
 	return (0);
